+++ neclib/core/data_type/parameters.py
@@ -31,7 +31,7 @@ from astropy.coordinates import Angle
 
 from ..exceptions import NECSTAccessibilityWarning, NECSTParameterNameError
 from ..files import toml
-from ..formatting.html_repr import html_repr_of_dict
+from ..formatting.dict_to_html import html_repr_of_dict
 
 
 class Parameters:
