+++ neclib/core/data_type/rich_parameters.py
@@ -9,7 +9,7 @@ from tomlkit.items import Item
 
 from ..exceptions import NECSTAccessibilityWarning, NECSTParameterNameError
 from ..files import toml
-from ..formatting.html_repr import html_repr_of_dict
+from ..formatting.dict_to_html import html_repr_of_dict
 from .parameters import Parameters
 
 T_value = TypeVar("T_value")
